Move filter_fx_graph utility to _support
There is no need to import compiler or kernel codegen, which bring IREE, for a
five-line function that filters the node in an FX graph. This will enable us to
remove local imports in water.
