+++ wave_lang/kernel/wave/specialize.py
@@ -49,7 +49,7 @@ from .._support.tracing import CapturedTrace
 from .._support.location import CapturedLocation
 from .._support.indexing import IndexSequence, IndexSymbol, IndexExpr
 
-from ..compiler.kernel_codegen import filter_fx_graph
+from .._support.fx import filter_fx_graph
 from ..lang.global_symbols import *
 from ..ops.wave_ops import (
     MMA,
